Fix DeleteModCommand crashing from deleting a non-existing directory, which can be caused by other bugs.

+++ source/Reloaded.Mod.Launcher.Lib/Commands/Mod/DeleteModCommand.cs
@@ -30,6 +30,9 @@ public class DeleteModCommand : WithCanExecuteChanged, ICommand
             return;
         
         var directory = Path.GetDirectoryName(_modTuple!.Path) ?? throw new InvalidOperationException(Resources.ErrorFailedToGetDirectoryOfMod.Get());
-        Directory.Delete(directory, true);
+        if (Directory.Exists(directory))
+        {
+            Directory.Delete(directory, true);
+        }
     }
 }
